LayoutTests/inspector/css/modify-css-property-expected.txt

Testing that CSSStyleDeclaration update immediately after modifying its properties when it is not locked.


== Running test suite: ModifyCSSProperty
-- Running test case: Update value when CSSStyleDeclaration is not locked
PASS: "font-size" property value should update immediately.
PASS: Style declaration text should update immediately.

-- Running test case: Update value when CSSStyleDeclaration is locked
PASS: "font-size" property value should update immediately.
PASS: Style declaration text should update immediately.

-- Running test case: ModifyCSSProperty.PropertiesChangedEvent
PASS: Style declaration is unlocked.
PASS: "width" property value should update to "200px".
PASS: Inline style declaration text should update when not locked.

-- Running test case: Update inline style value when CSSStyleDeclaration locked and not locked
PASS: Style declaration text should update immediately.
PASS: Style declaration text should update immediately.
PASS: Style declaration is unlocked.
PASS: "width" property value should update to "128px".
PASS: Inline style declaration text should update when not locked.

-- Running test case: ModifyCSSProperty.ConsequentValueChanges
PASS: Style declaration text should contain most recent value.

-- Running test case: ModifyCSSProperty.CommentOutAndUncommentPropertyWithNewlines
PASS: Commented out property should be disabled.
PASS: Style declaration text should update immediately with uncommented property.
PASS: Uncommented property should be enabled.
PASS: Style declaration text should update immediately with commented out property.
PASS: Commented out property should be disabled.

-- Running test case: ModifyCSSProperty.CommentOutAndUncommentPropertyWithoutNewlines
PASS: Commented out property should be disabled.
PASS: Style declaration text should update immediately with uncommented property.
PASS: Uncommented property should be enabled.
PASS: Commented out property should be disabled.
PASS: Style declaration text should update immediately with commented out property.
PASS: Uncommented property should be enabled.
